SEQNG-468 Improve robustness of setting views
