+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/components/SeqexecUI.scala
@@ -4,16 +4,13 @@
 package seqexec.web.client.components
 
 import diode.ModelRO
-import diode.react.ModelProxy
-import diode.react.ReactPot._
 import cats.implicits._
 import cats.effect.IO
 import monocle.Prism
 import japgolly.scalajs.react.vdom.html_<^._
 import japgolly.scalajs.react.extra.router._
-import japgolly.scalajs.react.{Callback, ScalaComponent}
+import japgolly.scalajs.react.Callback
 import japgolly.scalajs.react.MonocleReact._
-import japgolly.scalajs.react.component.Scala.Unmounted
 import gem.Observation
 import gem.enum.Site
 import scala.scalajs.js.timers.SetTimeoutHandle
@@ -22,96 +19,10 @@ import seqexec.web.client.actions.WSConnect
 import seqexec.web.client.model.Pages._
 import seqexec.web.client.ModelOps._
 import seqexec.web.client.actions.{NavigateSilentTo, RequestSoundEcho}
-import seqexec.web.client.components.sequence.{HeadersSideBar, SequenceArea}
 import seqexec.model.enum.Instrument
-import seqexec.web.client.model.WebSocketConnection
-import web.client.style._
-
-object AppTitle {
-  final case class Props(site: Site, ws: ModelProxy[WebSocketConnection])
-
-  private val component = ScalaComponent.builder[Props]("SeqexecUI")
-    .stateless
-    .render_P(p =>
-      <.div(
-        ^.cls := "ui row",
-        SeqexecStyles.shorterRow,
-        SeqexecStyles.notInMobile,
-        <.h4(
-          ^.cls := "ui horizontal divider header",
-          s"Seqexec ${p.site.shortName}",
-          p.ws().ws.renderPending(_ =>
-            <.div(
-              SeqexecStyles.errorText,
-              SeqexecStyles.blinking,
-              "Connection lost"
-            )
-          )
-        )
-      )
-    ).build
-
-  def apply(site: Site, ws: ModelProxy[WebSocketConnection]): Unmounted[Props, Unit, Unit] = component(Props(site, ws))
-}
-
-object SeqexecMain {
-  final case class Props(site: Site, ctl: RouterCtl[SeqexecPages])
-
-  private val lbConnect = SeqexecCircuit.connect(_.uiModel.loginBox)
-  private val logConnect = SeqexecCircuit.connect(_.uiModel.globalLog)
-  private val resourcesBusyConnect = SeqexecCircuit.connect(_.uiModel.resourceConflict)
-  private val headerSideBarConnect = SeqexecCircuit.connect(SeqexecCircuit.headerSideBarReader)
-  private val wsConnect = SeqexecCircuit.connect(_.ws)
-
-  private val component = ScalaComponent.builder[Props]("SeqexecUI")
-    .stateless
-    .render_P(p =>
-      <.div(
-        <.div(
-          ^.cls := "ui horizontally padded grid",
-          <.div(
-            ^.cls := "ui row",
-            SeqexecStyles.shorterRow
-          ),
-          wsConnect(ws => AppTitle(p.site, ws)),
-          <.div(
-            ^.cls := "ui row",
-            SeqexecStyles.shorterRow,
-            SeqexecStyles.queueAreaRow,
-            <.div(
-              ^.cls := "sixteen wide mobile ten wide tablet ten wide computer column",
-              SeqexecStyles.queueArea,
-              QueueTableSection(p.ctl)
-            ),
-            <.div(
-              ^.cls := "six wide column tablet computer only",
-              SeqexecStyles.headerSideBarArea,
-              headerSideBarConnect(HeadersSideBar.apply)
-            )
-          ),
-          <.div(
-            ^.cls := "ui row",
-            SeqexecStyles.shorterRow,
-            SequenceArea(SequenceArea.Props(p.ctl, p.site))
-          ),
-          <.div(
-            ^.cls := "ui row",
-            // Add margin to avoid covering the footer
-            SeqexecStyles.logArea,
-            logConnect(l => LogArea(p.site, l))
-          )
-        ),
-        lbConnect(LoginBox.apply),
-        resourcesBusyConnect(ResourcesBox.apply),
-        Footer(p.site)
-      )
-    ).build
-
-  def apply(site: Site, ctl: RouterCtl[SeqexecPages]): Unmounted[Props, Unit, Unit] = component(Props(site, ctl))
-}
 
 /**
-  * Top level UI component
+  * UI Router
   */
 object SeqexecUI {
   def pageTitle(site: Site)(p: SeqexecPages): String = p match {
