+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/model/actions.scala
@@ -35,9 +35,9 @@ object actions {
   case object Logout extends Action
 
   // Action to select a sequence for display
-  final case class SelectIdToDisplay(id: Observation.Id) extends Action
+  final case class SelectIdToDisplay(i: Instrument, id: Observation.Id) extends Action
   final case class SelectSequenceConfig(id: Observation.Id, step: StepId) extends Action
-  final case class SelectSequencePreview(id: Observation.Id, step: StepId) extends Action
+  final case class SelectSequencePreview(i: Instrument, id: Observation.Id, step: StepId) extends Action
   case object SelectEmptyPreview extends Action
 
   // Actions related to executing sequences
@@ -67,7 +67,7 @@ object actions {
   final case class RunObsPauseFailed(s: Observation.Id) extends Action
   final case class RunObsResumeFailed(s: Observation.Id) extends Action
 
-  final case class ShowStepConfig(id: Observation.Id, step: Int, isPreview: Boolean) extends Action
+  final case class ShowStepConfig(i: Instrument, id: Observation.Id, step: Int, isPreview: Boolean) extends Action
   final case class HideStepConfig(i: Instrument) extends Action
   final case class RememberCompleted(s: SequenceView) extends Action
